refactor(drivers): extract isUUIDLike to shared helper with tests
